+++ Gruntfile.js
@@ -1,3 +1,9 @@
+/**
+ * Gruntfile.js
+ *
+ * Run 'grunt' in shell to compile javascript and css files
+ *
+ */
 module.exports = function(grunt) {
 
   grunt.initConfig({
@@ -23,13 +29,28 @@ module.exports = function(grunt) {
           'css/main.min.css': ['css/main.css', 'css/tagcloud.css', 'css/jquery.rating.css']
         }
       }
+    },
+    shell: {
+      builddoc: {
+        command: 'cd doc; make html'
+      },
+      buildapi: {
+        command: 'phpdoc run -d inc/classes -t doc/api'
+      },
+      runtests: {
+        command: 'php vendor/bin/codecept run --skip acceptance'
+      }
     }
   });
 
   grunt.loadNpmTasks('grunt-contrib-uglify');
   grunt.loadNpmTasks('grunt-contrib-watch');
   grunt.loadNpmTasks('grunt-contrib-cssmin');
+  grunt.loadNpmTasks('grunt-shell');
 
   grunt.registerTask('default', ['uglify', 'cssmin']);
+  grunt.registerTask('doc', 'shell:builddoc');
+  grunt.registerTask('api', 'shell:buildapi');
+  grunt.registerTask('test', 'shell:runtests');
 
 };